requirements.txt
APScheduler==3.11.0
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.0
idna==3.10
iniconfig==2.1.0
kiwisolver==1.4.8
matplotlib==3.10.3
mpmath==1.3.0
mypy_extensions==1.1.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
setuptools==80.8.0
six==1.17.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
wheel==0.45.1
PyYAML
